Create DSL-Assignment-4
Assignment 4
Write a Python program that computes the net amount of a bank account based a transaction log from console input. The transaction log format is shown as following: D 100 W 200(Withdrawal is not allowed if balance is going negative. Write functions for withdraw and deposit)D means while W means withdrawal. Suppose the following input is supplied to the program: D 300, D 300, W 200, D 100 Then, the output should be: 500